bitset iter done

+++ libs/base/include/base/bitset.h
@@ -7,7 +7,7 @@
 #include <stdarg.h>
 
 struct bitset {
-    unsigned nwords;
+    unsigned short nwords;
     size_t words[];
 };
